+++ angular/lib/src/source_gen/template_compiler/find_components.dart
@@ -524,7 +524,7 @@ class ComponentVisitor
       preserveWhitespace: coerceBool(
         component,
         'preserveWhitespace',
-        defaultTo: true,
+        defaultTo: false,
       ),
     );
   }